requirements.txt
python-docx>=0.8.11
graphviz>=0.20
matplotlib>=3.5
numpy>=1.20
streamlit>=1.28
 